App useEffect => loader in RootLayout
